requirements.txt
schwabdev>=2.5.0
requests>=2.25.0
websockets>=10.0
PyYAML>=6.0